backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.2
python-dotenv==1.2.1
bcrypt==5.0.0
pillow==11.3.0
waitress==3.0.2
gunicorn==21.2.0
